backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
supabase==2.0.2
psycopg2-binary

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI and ML
requests==2.31.0
openai==1.3.7

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7

# Background tasks
celery==5.3.4
redis==5.0.1

# Data validation and serialization
pydantic==2.11.7
pydantic[email]==2.11.7

# HTTP client
httpx>=0.24.0,<0.25.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0

# Optional: For better performance
orjson==3.9.10